dashboard-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
boto3==1.34.0
pyjwt[crypto]==2.8.0
cryptography==42.0.5
pydantic==2.7.1
python-dotenv==1.0.1
